requirements.txt
Django>=2.0
django-admin
djangorestframework
django-filter
tushare
pandas>=0.20.0
urllib3
requests
easyutils
quantaxis>1.0.33
selenium
lxml<=4.1.1
bs4
